Add basic priority-weighted occupancy
This doesn't make the test pass, because of exactly what I theorized in https://github.com/dask/distributed/issues/5253#issuecomment-915411871: once one of the proper downstream tasks has been assigned to a worker, the next tasks get sent somewhere else, because they can cut to position 0 in the line elsewhere instead of position 1 on that worker. We'd need the communication cost to be higher in order to beat this.

Also this metric is still slightly off, because it's acting as though workers are single-threaded. In reality position 0 and 1 in the queue could maybe both run at (nearly) the same time. Also, using average `ws.occupancy` means if the root tasks are slower than the downstream tasks, we're making `stack_time` too high (when i==1, `stack_time` gets counted as the runtime for a `slowinc` task, not the one fast `inc` task that's actually ahead in the priority queue).